xen/arm: introduce 1:1 mapping for domUs
In some cases it is desirable to map domU memory 1:1 (guest physical ==
physical.) For instance, because we want to assign a device to the domU
but the IOMMU is not present or cannot be used. In these cases, other
mechanisms should be used for DMA protection, e.g. a MPU.

This patch introduces a new device tree option for dom0less guests to
request a domain to be directly mapped. It also specifies the memory
ranges. This patch documents the new attribute and parses it at boot
time. Finally the patch sets the new direct_map flag for DomU domains.

Given that with direct-map the IOMMU can be used but it is not required,
avoid setting XEN_DOMCTL_CDF_iommu when the IOMMU is disabled and
direct_map is requested.

Signed-off-by: Stefano Stabellini <stefano.stabellini@xilinx.com>
Changes summary: Ported to apply on 'develop-4.14' branch
Cherry-pick-repository: https://github.com/Xilinx/xen.git
(cherry picked from commit 855e290b579bdcf6d2442b41d7754c2b09d753d1)
Issue: ALB-6261
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>