uart: Fix getc/putc/tx_ready implementation logic
Separated the implementation logic in getc(), putc() and tx_ready()
functions to match the LINFlex UART communication protocol in
Buffer and FIFO modes.

In TX Buffer mode (UARTCR.TFBM == 0), the hardware logic does not
match well with the Xen design of uart_driver ops. Before calling
putc(), Xen calls tx_ready() to check if the console is ready to
print the character. However, our Buffer mode protocol relies on
sending the character without any apriori checks, then waiting
until the Data Transmission Completed Flag (UARTSR.DTF) is set
by the hardware to 1 and, afterwards, clearing it in software.

Therefore, tx_ready() is a 'dummy' function in this case, returning
the buffer size whenever called. The UARTSR.DTF check and clear are
done after sending the character, in putc().

Removed an unnecessary check of UARTSR.RMB in getc() function and
refactored the code to match the same general style.

Issue: ALB-5666
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>