arch/arm: irq: Add platform_get_irq_byname() implementation
Moved implementation for the function which parses the IRQs of a DT
node by the "interrupt-names" property from the SMMU-v3 driver
to the IRQ core code and made it non-static to be used as helper.

Also changed it to receive a "struct dt_device_node*" as parameter,
like its counterpart, platform_get_irq(). Updated its usage inside
the SMMU-v3 driver accordingly.

Issue: ALB-9628
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>