arm: Update guest memory map to reflect s32cc resources
This is needed for the xl toolstack which will try to allocate
a buffer grant region when dom0less PV drivers are used. The
region will be allocated by default at GUEST_MAGIC_BASE, and
it will be used as a shmem channel between Dom0 and the Dom0less
DomU, such that the PV FE/BE drivers can communicate.

Also update the DRAM ranges while at it.

Issue: ALB-9508
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>