build: omit "source" symlink when building hypervisor in-tree
This symlink is getting in the way of using e.g. "find" on the xen/
subtree, and it isn't really needed when not building out-of-tree:
The one use that there was can easily be avoided.

Signed-off-by: Jan Beulich <jbeulich@suse.com>
Reviewed-by: Anthony PERARD <anthony.perard@citrix.com>

This is a cherry-pick of:
commit 96467d77b4e4ab032145b39eee86e463871ade
(build: omit "source" symlink when building hypervisor in-tree")

Issue: ALB-10427

Signed-off-by: Georgiana Ortila <georgiana.ortila@nxp.com>