CI: Fix the github actions Coverity run
The Install step is failing with:

  Package python-dev is not available, but is referred to by another package.
  This may mean that the package is missing, has been obsoleted, or
  E: Package 'python-dev' has no installation candidate
  is only available from another source
  However the following packages replace it:
    python2-dev python2 python-dev-is-python3

which is a side effect of ubuntu-latest having changed from 20.04 to 22.04.

Drop python-dev as python3-dev is already present, and used a fixed runs-on to
avoid such problems in the future.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
Acked-by: Stefano Stabellini <sstabellini@kernel.org>
Reviewed-by: Roger Pau Monné <roger.pau@citrix.com>

This is a cherry-pick of:
commit 606bb708e13d64596b2d68ac1de78d142d682e20
("CI: Fix the github actions Coverity run")

Issue: ALB-10322
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>